Move custom date forms into handler
