+++ app/forms/field_handlers/year_date_handler.py
@@ -1,6 +1,6 @@
-from wtforms import StringField
+from werkzeug.utils import cached_property
+from wtforms import StringField, Form
 
-from app.forms.date_form import DateForm
 from app.forms.field_handlers.date_handler import DateHandler
 from app.forms.validators import SingleDatePeriodCheck
 
@@ -9,8 +9,17 @@ class YearDateHandler(DateHandler):
     DATE_FORMAT = 'yyyy'
 
     def get_form_class(self):
-        class YearDateForm(DateForm):
+        class YearDateForm(Form):
             year = StringField(validators=self.validators)
+
+            @cached_property
+            def data(self):
+                data = super().data
+
+                try:
+                    return '{year:04d}'.format(year=int(data['year']))
+                except (TypeError, ValueError):
+                    return None
 
         return YearDateForm
